Remove Abstract base class - no longer needed after V1 runtime deleted
