+++ core/mybatis-generator-core/src/main/java/org/mybatis/generator/runtime/dynamic/sql/DynamicSqlMapperGenerator.java
@@ -15,14 +15,22 @@
  */
 package org.mybatis.generator.runtime.dynamic.sql;
 
+import static org.mybatis.generator.internal.util.StringUtility.stringHasValue;
 import static org.mybatis.generator.internal.util.messages.Messages.getString;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import org.mybatis.generator.api.dom.java.CompilationUnit;
+import org.mybatis.generator.api.dom.java.FullyQualifiedJavaType;
 import org.mybatis.generator.api.dom.java.Interface;
+import org.mybatis.generator.api.dom.java.JavaVisibility;
 import org.mybatis.generator.api.dom.java.TopLevelClass;
+import org.mybatis.generator.codegen.AbstractJavaClientGenerator;
+import org.mybatis.generator.codegen.AbstractXmlGenerator;
+import org.mybatis.generator.config.PropertyRegistry;
+import org.mybatis.generator.internal.util.JavaBeansUtil;
+import org.mybatis.generator.runtime.dynamic.sql.elements.AbstractMethodGenerator;
 import org.mybatis.generator.runtime.dynamic.sql.elements.BasicCountMethodGenerator;
 import org.mybatis.generator.runtime.dynamic.sql.elements.BasicDeleteMethodGenerator;
 import org.mybatis.generator.runtime.dynamic.sql.elements.BasicInsertMethodGenerator;
@@ -32,6 +40,7 @@ import org.mybatis.generator.runtime.dynamic.sql.elements.BasicSelectOneMethodGe
 import org.mybatis.generator.runtime.dynamic.sql.elements.BasicUpdateMethodGenerator;
 import org.mybatis.generator.runtime.dynamic.sql.elements.DeleteByPrimaryKeyMethodGenerator;
 import org.mybatis.generator.runtime.dynamic.sql.elements.FieldAndImports;
+import org.mybatis.generator.runtime.dynamic.sql.elements.FragmentGenerator;
 import org.mybatis.generator.runtime.dynamic.sql.elements.GeneralCountMethodGenerator;
 import org.mybatis.generator.runtime.dynamic.sql.elements.GeneralDeleteMethodGenerator;
 import org.mybatis.generator.runtime.dynamic.sql.elements.GeneralSelectDistinctMethodGenerator;
@@ -41,6 +50,7 @@ import org.mybatis.generator.runtime.dynamic.sql.elements.GeneralUpdateMethodGen
 import org.mybatis.generator.runtime.dynamic.sql.elements.InsertMethodGenerator;
 import org.mybatis.generator.runtime.dynamic.sql.elements.InsertMultipleMethodGenerator;
 import org.mybatis.generator.runtime.dynamic.sql.elements.InsertSelectiveMethodGenerator;
+import org.mybatis.generator.runtime.dynamic.sql.elements.MethodAndImports;
 import org.mybatis.generator.runtime.dynamic.sql.elements.SelectByPrimaryKeyMethodGenerator;
 import org.mybatis.generator.runtime.dynamic.sql.elements.SelectListGenerator;
 import org.mybatis.generator.runtime.dynamic.sql.elements.UpdateAllColumnsMethodGenerator;
@@ -48,10 +58,21 @@ import org.mybatis.generator.runtime.dynamic.sql.elements.UpdateByPrimaryKeyMeth
 import org.mybatis.generator.runtime.dynamic.sql.elements.UpdateByPrimaryKeySelectiveMethodGenerator;
 import org.mybatis.generator.runtime.dynamic.sql.elements.UpdateSelectiveColumnsMethodGenerator;
 
-public class DynamicSqlMapperGenerator extends AbstractDynamicSqlMapperGenerator {
+public class DynamicSqlMapperGenerator extends AbstractJavaClientGenerator {
+
+    // record type for insert, select, update
+    protected FullyQualifiedJavaType recordType;
+
+    // id to use for the common result map
+    protected String resultMapId;
+
+    // name of the field containing the table in the support class
+    protected String tableFieldName;
+
+    protected FragmentGenerator fragmentGenerator;
 
     public DynamicSqlMapperGenerator(String project) {
-        super(project);
+        super(project, false);
     }
 
     @Override
@@ -103,6 +124,47 @@ public class DynamicSqlMapperGenerator extends AbstractDynamicSqlMapperGenerator
         }
 
         return answer;
+    }
+
+    protected void preCalculate() {
+        recordType = new FullyQualifiedJavaType(introspectedTable.getBaseRecordType());
+        resultMapId = recordType.getShortNameWithoutTypeArguments() + "Result"; //$NON-NLS-1$
+        tableFieldName =
+                JavaBeansUtil.getValidPropertyName(introspectedTable.getFullyQualifiedTable().getDomainObjectName());
+        fragmentGenerator = new FragmentGenerator.Builder()
+                .withIntrospectedTable(introspectedTable)
+                .withResultMapId(resultMapId)
+                .withTableFieldName(tableFieldName)
+                .build();
+    }
+
+    protected Interface createBasicInterface() {
+        FullyQualifiedJavaType type = new FullyQualifiedJavaType(
+                introspectedTable.getMyBatis3JavaMapperType());
+        Interface interfaze = new Interface(type);
+        interfaze.setVisibility(JavaVisibility.PUBLIC);
+        context.getCommentGenerator().addJavaFileComment(interfaze);
+        interfaze.addImportedType(new FullyQualifiedJavaType("org.apache.ibatis.annotations.Mapper")); //$NON-NLS-1$
+        interfaze.addAnnotation("@Mapper"); //$NON-NLS-1$
+
+        String rootInterface = introspectedTable
+                .getTableConfigurationProperty(PropertyRegistry.ANY_ROOT_INTERFACE);
+        if (!stringHasValue(rootInterface)) {
+            rootInterface = context.getJavaClientGeneratorConfiguration()
+                    .getProperty(PropertyRegistry.ANY_ROOT_INTERFACE);
+        }
+
+        if (stringHasValue(rootInterface)) {
+            FullyQualifiedJavaType fqjt = new FullyQualifiedJavaType(rootInterface);
+            interfaze.addSuperInterface(fqjt);
+            interfaze.addImportedType(fqjt);
+        }
+        return interfaze;
+    }
+
+    protected TopLevelClass getSupportClass() {
+        return DynamicSqlSupportClassGenerator.of(
+                introspectedTable, context.getCommentGenerator(), warnings).generate();
     }
 
     protected void addInsertOneMethod(Interface interfaze) {
@@ -365,5 +427,21 @@ public class DynamicSqlMapperGenerator extends AbstractDynamicSqlMapperGenerator
                 .build();
 
         generate(interfaze, generator);
+    }
+
+    protected boolean generate(Interface interfaze, AbstractMethodGenerator generator) {
+        MethodAndImports mi = generator.generateMethodAndImports();
+        if (mi != null && generator.callPlugins(mi.getMethod(), interfaze)) {
+            interfaze.addMethod(mi.getMethod());
+            interfaze.addImportedTypes(mi.getImports());
+            interfaze.addStaticImports(mi.getStaticImports());
+            return true;
+        }
+        return false;
+    }
+
+    @Override
+    public AbstractXmlGenerator getMatchedXMLGenerator() {
+        return null;
     }
 }
